Chemical sample at point is now net of attractant and repellent.
Mimicking diffusion, the Chemical concentration at a point is the net value of all the gradients of every chemical in the Agar.

+++ Assets/Scripts/Chemical.cs
@@ -25,9 +25,9 @@ public class Chemical : MonoBehaviour {
         this.concentration = concentration;
     }
 
-    public float getConcentration(Vector3 position)
+    public int getConcentration(Vector3 position)
     {
-        return this.concentration * (transform.localScale.magnitude / Vector3.Distance(origin, position));
+        return (int) (this.concentration * (transform.localScale.magnitude / Vector3.Distance(origin, position)));
     }
 
     public void setEcoliReaction(BacteriaReaction reaction)
